revert finstSteps
